use component bow knockback
Vanilla represents Punch as the bow weapon's KNOCKBACK value effect with an arrow direct-attacker condition. The local bow path still read the old integer Punch enchantment level and stored the resulting arrow knockback as an int. Default Punch mostly worked, but custom or fractional component-driven bow knockback values could not be represented.

The fix reads the bow's KNOCKBACK value effect when creating the arrow and changes arrow knockback storage to a double. The existing arrow knockback application formula is unchanged, so default Punch levels keep their vanilla strength while component-based values can now flow through.

+++ src/main/java/io/github/togar2/pvp/entity/projectile/AbstractArrow.java
@@ -39,7 +39,7 @@ public abstract class AbstractArrow extends CustomEntityProjectile {
 	protected int ticks;
 	private double baseDamage = ARROW_BASE_DAMAGE;
 	private double damageBonus = 0.0;
-	private int knockback;
+	private double knockback;
 	private ItemStack weaponItem = ItemStack.AIR;
 	private SoundEvent soundEvent = getDefaultSound();
 
@@ -321,11 +321,11 @@ public abstract class AbstractArrow extends CustomEntityProjectile {
 		return 0.6;
 	}
 
-	public int getKnockback() {
+	public double getKnockback() {
 		return knockback;
 	}
 
-	public void setKnockback(int knockback) {
+	public void setKnockback(double knockback) {
 		this.knockback = knockback;
 	}
